Create-ReadMe.Md
The end of one life and the beginning of MY life.... I'm not here to make a mark on this world. I'm here to fix the marks that shouldn't be on it. ... Not For Hire Inc.. we don't do a damn thing because we "have" to...  we do it because we chose to....  and the only pay we accept is paying it forward.